### 2020-05-07 #42 @ev-br
Ten years challenge: fermions at unitarity
**Original article:** https://journals.aps.org/prl/abstract/10.1103/PhysRevLett.96.160402
Preprint, ungated: https://arxiv.org/abs/cond-mat/0602224

**Scientific domain:** Physics
**Programming language:** Fortran / Python

This is paper 36 in https://github.com/ReScience/ten-years/issues/1
