### 2020-05-19 #42 @pdebuyl
@jochym could you review the submission here? If this is your first review for ReScience, I will guide you through the process.
