### 2020-05-27 #42 @pdebuyl
@dombrno could you review the submission here? If this is your first review for ReScience, I will guide you through the process.
