### 2020-06-23 #42 @pdebuyl
Hi @vahtras

We have reviewer guidelines here: https://rescience.github.io/edit/

The most specific part of the review, in relation to ReScience, is to actually run the code and verify the claims of reproduction that are stated in the article.

Specifically for the ten-year challenge
- We only require one reviewer.
- The requirement to have a readable and reusable code are relaxed a bit. It would make little sense for the authors of this issue to perform a full modernization of their code.
- The author should provide in the article their reflection on the longevity, quality and retrievable character of their old code and of the corresponding environment (language, platform-specific code, proprietary tools, etc).
