### 2020-06-24 #42 @pdebuyl
How many core-hours ? (or core-days / node-days depending on the hardware you have used)?
