### 2020-06-24 #42 @ev-br
The largest runs in the supplement repository are some 120 CPU hours on 24 cores. Multiply it by about 1.5-2 for thermalization.
Smaller system sizes are *much* faster, some 4-10 cpu hours each, if a partial verification is OK.  (rerun small system sizes on the reviewer's machine, rerun the fits with a mix of reviewer's data and my data or somesuch).
