### 2020-06-26 #42 @pdebuyl
Well, this seems indeed costly. @vahtras do you have the resources for the small systems verification? (4-10 cpu hours each) ? I could execute that if necessary.
