### 2020-07-30 #42 @pdebuyl
Hi all,

I see only now that the review process is frozen here. @vahtras what do you think of the computational requirements?
