### 2020-08-11 #42 @vahtras
I apologize for the delay, will have some time for a look now.

On Thu, Jul 30, 2020 at 10:10 PM Pierre de Buyl <notifications@github.com>
wrote:

> Hi all,
>
> I see only now that the review process is frozen here. @vahtras
> <https://github.com/vahtras> what do you think of the computational
> requirements?
>
