### 2020-09-09 #42 @vahtras
Yes, finally. First question: the makefile has been hardcoded for Intel compilers. Does it build with GNU?
