### 2020-09-09 #42 @ev-br
It certainly does. https://github.com/ev-br/10yr_repro_challenge_35/commit/837038f2461ee5e17da895388bda935749918e0c is the relevant makefile.

I then simply commented it out when transferring to cluster/intel compiler instead of adding platform detection (more brittle stuff to debug ten years down the line)

The change from gnu on a laptop to intel on cluster is here:
https://github.com/ev-br/10yr_repro_challenge_35/commit/3603020b726cd6eab0db18d03c687d9237d757d7

EDIT:the switch from gnu to intel is
https://github.com/ev-br/10yr_repro_challenge_35/commit/acc29743ec4f0d49da013c48db2ef36e0a48cb50
