### 2021-06-23 #42 @vahtras
please find another reviewer

On Tue, Jun 22, 2021 at 8:56 AM Pierre de Buyl ***@***.***>
wrote:

> Hi @rougier <https://github.com/rougier> sorry about this. @vahtras
> <https://github.com/vahtras> is this still doable for you? In the
> meantime, I will try to find another reviewer.
>
