### 2021-07-22 #42 @pdebuyl
Hi @rougier unfortunately I "lost" my candidate reviewer. In all fairness (and despite the already long delay), I'd prefer to wait mid-august to look again the mid-summer is a really bad time to find reviewers!

PS: I'll also consider reviewing the paper and have Konrad or someone else be editor (august as well I am not in office)
