### 2021-07-27 #42 @rougier
Ok. Note that according to our process, you can serve as 2nd reviewer while being the editor.
