### 2025-03-07 #42 @ev-br
Out of curiosity: "close as completed" means it's published or that it's rejected?
