### 2025-03-23 #42 @ev-br
Thanks Nicolas! Yes, I am still interested. Would appreciate if you proceed.

Since I did the work already back then, for me it's strictly better that it's reviewed with a plausible path to being published, than that it is simply gone :-).

On a side note, overall the story here so far can be viewed as a demonstration that a volunteer open-source model is not a full replacement for"traditional publishers", with all their sins. Not that it's specific to scientific related publishing, though.
