### 2025-05-06 #42 @khinsen
@ev-br After a first pass through the manuscript (nice to read, detailed comments later), I took a look at the code and decided to try running the small test case myself. Here is what I did and what happened - do you see some obvious mistake I made?

```
guix shell gfortran-toolchain lapack -- \
   gfortran -llapack -lblas det_n2.f event_mod.f rndm_mumbers.f fermi-hubbard.f
cp par_tst infile
./a.out
```

Outcome:
```
At line 436 of file fermi-hubbard.f (unit = 1, file = '!')
Fortran runtime error: End of file

Error termination. Backtrace:
#0  0x7f60f2ecc682 in ???
#1  0x7f60f2ecd149 in ???
#2  0x7f60f30f36db in ???
#3  0x7f60f30ed104 in ???
#4  0x7f60f30edcb9 in ???
#5  0x406a0d in MAIN__
#6  0x41ba83 in main
```
