### 2025-05-06 #42 @ev-br
Thanks for taking a look @khinsen !
The culprit is
`$ cp par_tst infile`
both files need to be present.

```
$ gfortran det_n2.f event_mod.f rndm_mumbers.f fermi-hubbard.f -llapack -lblas
$ cat infile
1
1 par_tst
$ ls par*
par_tst
$ ./a.out
```

Also there will be no output in terminal; all output goes to `out_tst_1`. The program will terminate after the specified time limit:

```
$ tail par_tst
1.d0      ! step for measuring
1.7       ! time limit, hrs          <----- here
------
```

---------------

If this workflow sounds obnoxious to you, I totally agree! One of the first things we did with that code was to replace all MPI with shell scripting, but for this reproduction experiment I stick to the original workflow from the original paper.
