### 2025-05-09 #42 @khinsen
Thanks @ev-br ! It looks like I was looking for something more complicated. I started from the unmodified repository, just ran `a.out`, and it ran without any complaint.

However, the README says near the end:
```
gim.dat --- ct3d_tst.dat file from your simulation
            must match this one
Vsmall.dat --- nmnm_tst.dat must match this one
```
Neither file matches for me, not even in the number of lines, let alone the values.

Example:
```
$ head gim.dat
  -0.2925000000000000        7.1858851844498677E-003
  -0.2775000000000000        2.1475228631994816E-002
  -0.2625000000000000        3.3147014723346424E-002
  -0.2475000000000000        4.9523300597582951E-002
  -0.2325000000000000        6.3246243769281874E-002
  -0.2175000000000000        7.8902265781143432E-002
  -0.2025000000000000        9.4774608201243407E-002
  -0.1875000000000000        0.1114615698538545
  -0.1725000000000000        0.1316859196939726
  -0.1575000000000000        0.1479454676031685
$ head ct3d_tst_1.dat
 -0.28999999999999998        9.1143942058506001E-003
 -0.26999999999999996        2.7771096455593988E-002
 -0.25000000000000000        4.6406376286320286E-002
 -0.22999999999999998        6.6132076590356845E-002
 -0.20999999999999999        8.6604929575525819E-002
 -0.19000000000000000       0.10817597680630522
 -0.16999999999999998       0.13155242792684838
 -0.14999999999999999       0.15676221659046960
 -0.13000000000000000       0.18564518648366518
 -0.11000000000000000       0.21543432266493209
```

I hope that I didn't break any fermions!
