### 2025-05-09 #42 @ev-br
Ouch. You certainly did not break any fermions, but the guy who wrote this "must match" sentence in 2006 was being a little gung-ho optimistic. The phrase is really from the 2006 demo (there's `The rest of this file is the README from 2006, verbatim.` line in the README, what's above it was written in 2020; what's below is from 2006).

First, the original readme is unversioned, and to be honest I've no idea if the "demo" data files were produced by a single run with the provided parameter file (which sets the time limit of a couple of hours IIUC) or a longer (overnight?) run.

Second, the "must match" is at best "when plotted, the overall shape should asymptotically be similar to".
Let me explain.
These data files dump sampled distributions of rather technical parameters generated by a Monte Carlo simulation. As such, they are at best useful for qualitative checks, but specific numbers have no meaning.

In detail:

- Vsmall.dat / nmnm.dat : the simulation samples Feynman diagrams, and this file dumps the distribution of the diagram order. Thus, it should converge to a Gaussian shape. The center of the Gaussian roughly scales as $\sim \beta U N $, where $\beta$ is the inverse temperature,  $U$ is the Hubbard coupling constant and $N$ is the number of particles. So if you double the `beta` parameter in the `par_` file, or tune the chemical potential, (`mu` in the parameter file) to double the number of particles (which is related to `dens` in the output files), the Gaussian peak should shift to the right by a factor of roughly two.

All this is in the asymptotics of course. Requiring that two sampled distributions "must match" is really bad wording (I take full responsibility for what a grad student with my name wrote though).

Checking the overall shape of the distribution is actually a useful visual check of convergence of simulations. IIRC I did some analysis of this in the paper and the accompanying notebook --- for the larger reproduction simulation I did for the paper.

Apparently, it would have been useful to do the analysis of the toy problem from the README. But I should not pre-empt reviewer's comments :-).

- the other file, `gim.dat`, is harder to make sense of. What it dumps is a distribution of the four-point pair-pair correlation function, in the imaginary time, integrated over the two coordinates and one time variable. Sampled on some fixed grid.
So, specific numbers in that file have no meaning whatsoever.
What do fermions have to say about the overall shape? There should be a jump at zero, sure. (because it's something like
$\langle c_\uparrow c_\downarrow (0) c_\uparrow^\dagger c_\downarrow^\dagger (\tau)  \rangle$, which jumps at $\tau=0$). It does not need to be symmetric around zero. It should probably decay with $\tau$. It should asymptotically converge to some well defined shape, not jump around. That's pretty much it, I'd think.
