### 2025-05-14 #42 @khinsen
Thanks @ev-br ! Much of what you just wrote would make a nice addition to the 2020 part of the README. After all, the goal of the whole reproduction exercise is to figure out what should have been done better in 2006, so these additional comments look valuable.
