### 2025-05-21 #42 @khinsen
@ev-br @pdebuyl @rougier Here comes my review!

This is a nicely told reproduction story with a clear message: if you use stable languages such as Fortran, you don't have to worry that much about computational environments, but you still must document your workflow carefully, and keep all the input (code, input parameters, command line incantations) under version control so that you know what you actually did in a production run.

The paper also makes an important technical point: workflow automation adds dependencies, which have a good chance of becoming liabilities in the future. Human-readable workflow documentation is less precise but more durable. I hope that one day we will have workflow tools that ensure automation while keeping the workflow human-readable, but as of today, that's not the case, so scientists should be aware of the tension.

I submitted a pull request to the article that fixes some typos, but other than that, it's just fine. Very readable, very complete.

The only (optional!) improvement I could suggest is some additions to the README that comes with the code, as suggested above.
