### 2025-05-23 #42 @khinsen
I just saw this post that supports @ev-br's admonition about workflow tooling: https://rollenspiel.social/@ArneBab/114555972996531535
