### 2025-05-25 #42 @ev-br
Thank you very much for the review @khinsen !
It is a great reality check to see things which I took for granted, but were in fact, plain misleading.

Also thank you for the PR with typo fixes. I merged in the PR (did not update the pdf just yet), and added some discussion + links to the README, as you suggested.

What would be the next step?

As for human-readable workflow documentation---indeed, I do believe it is crucial. Reproducibility is as at least as much a human problem as it is a technical problem, so a purely technical solution cannot be sufficient.

As a personal aside, one of the funding sources we worked with required very verbose reports in addition to peer-reviewed publications (on the order of ~50 pages of prose per one or two 4-page PRL papers). I used to grumble at this requirement, and it's a hassle to prepare, but in the end these detailed reports prove useful for sharing assumed knowledge with new people in the lab and with outside colleagues.
