### 2025-05-26 #42 @rougier
If @khinsen is satisifed with correction/fixes, I think we can proceed with publication (5 years later)...
