### 2025-05-27 #42 @khinsen
@rougier Yes, I am happy with the corrections. Please proceed with publication. We can then re-review this in five years to see if the review is reproducible ;-)
