### 2025-05-27 #42 @khinsen
@ev-br It's also why a thesis is often more useful than the paper(s) derived from it. Details are always helpful, even if theoretically redundant.
